pyproject.toml
[project]
name = "nutri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cors>=1.0.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "npm>=0.1.1",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "streamlit-option-menu>=0.4.0",
    "stripe>=12.3.0",
]
